refactor: move `_hooks` from `_utils` to `_collections`
