+++ src/site/_collections/_hooks/authors.js
