+++ src/site/content/en/authors/individual.11tydata.js
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-const {individual} = require('../../../_utils/_hooks/authors');
+const {individual} = require('../../../_collections/_hooks/authors');
 
 module.exports = {
   pagination: {
